requirements.txt
PyGithub
requests
argparse
pycurl
clang
spacy
numpy
matplotlib
sklearn